Make SearchCluster.TopKEstimator a top level class.
